Refactor ArcHTTPResolver to support both file and HTTP and rename it to RewriteLocationResolver
